feat: add auth

+++ src/controllers/cards.ts
@@ -1,10 +1,11 @@
 import { RequestHandler } from 'express'
 import Card from '../models/card'
-import type { UserRequest } from '../types'
+import type { AuthRequest } from '../types'
 import NotFoundError from '../errors/not-found-error'
 import { sendResponse } from '../helper'
+import ForbiddenError from '../errors/forbidden-error'
 
-export const createCard: RequestHandler = (req: UserRequest, res, next) => {
+export const createCard: RequestHandler = (req: AuthRequest, res, next) => {
   const { name, link } = req.body
 
   Card.create({ name, link, owner: req.user?._id })
@@ -18,13 +19,19 @@ export const getCards: RequestHandler = (req, res, next) => {
     .catch(next)
 }
 
-export const deleteCards: RequestHandler = (req, res, next) => {
-  Card.findByIdAndDelete(req.params.cardId)
-    .then(() => sendResponse(res, 'Успешно удалено'))
+export const deleteCards: RequestHandler = (req: AuthRequest, res, next) => {
+  Card.findById(req.params.cardId)
+    .then((card) => {
+      if (!card) throw new NotFoundError('Карточка не найдена')
+      if (card.owner.toString() !== req.user?._id) throw new ForbiddenError('Нет прав на удаление')
+      Card.findByIdAndDelete(req.params.cardId)
+        .then(() => sendResponse(res, 'Успешно удалено'))
+        .catch(next)
+    })
     .catch(next)
 }
 
-export const addLike: RequestHandler = (req: UserRequest, res, next) => {
+export const addLike: RequestHandler = (req: AuthRequest, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user?._id } }, { new: true })
     .then((card) => {
       if (!card) throw new NotFoundError('Нет карточки с таким id')
@@ -33,7 +40,7 @@ export const addLike: RequestHandler = (req: UserRequest, res, next) => {
     .catch(next)
 }
 
-export const deleteLike: RequestHandler = (req: UserRequest, res, next) => {
+export const deleteLike: RequestHandler = (req: AuthRequest, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user?._id } }, { new: true })
     .then((card) => {
       if (!card) throw new NotFoundError('Нет карточки с таким id')
